requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
bcrypt>=4.0.0
ultralytics>=8.0.0
requests>=2.31.0
python-telegram-bot>=21.0
customtkinter>=5.2.0
pyperclip>=1.8.0